Add new guard to BlockView activation

+++ src/client/src/components/header/handlers/OnView.tsx
@@ -1,14 +1,14 @@
 import { ViewType, VIEW_TYPE } from "../../../models/project";
 import { changeFlowView } from "../../../redux/store/flow/actions";
 import { removeSecondaryNode } from "../../../redux/store/secondaryNode/actions";
-import { GetSelectedNode } from "../../../helpers";
+import { GetSelectedNode, IsAspectNode } from "../../../helpers";
 import { updateBlockElements } from "../../../modules/explorer/redux/actions";
 
 const OnView = (view: ViewType, dispatch: any) => {
   const selectedNode = GetSelectedNode();
 
-  // BlockView can only be opened when a node is selected
-  if (view === VIEW_TYPE.BLOCKVIEW && !selectedNode) return;
+  // BlockView can only be opened when a non-aspect node is selected
+  if (view === VIEW_TYPE.BLOCKVIEW && (!selectedNode || IsAspectNode(selectedNode))) return;
 
   dispatch(removeSecondaryNode());
   dispatch(updateBlockElements([]));
